requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.5.0
    # via pydantic
anthropic==0.29.0
    # via -r requirements.in
anyio==3.7.1
    # via
    #   anthropic
    #   httpx
    #   openai
cachetools==5.3.3
    # via google-auth
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
distro==1.8.0
    # via
    #   anthropic
    #   openai
filelock==3.15.4
    # via huggingface-hub
fsspec==2024.6.0
    # via huggingface-hub
google-ai-generativelanguage==0.6.5
    # via google-generativeai
google-api-core[grpc]==2.19.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.134.0
    # via google-generativeai
google-auth==2.30.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.7.0
    # via -r requirements.in
googleapis-common-protos==1.63.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.64.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.1
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.25.1
    # via
    #   anthropic
    #   openai
huggingface-hub==0.23.4
    # via tokenizers
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.4.2
    # via anthropic
openai==1.0.1
    # via -r requirements.in
packaging==24.1
    # via huggingface-hub
proto-plus==1.24.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.4.2
    # via
    #   -r requirements.in
    #   anthropic
    #   google-generativeai
    #   openai
pydantic-core==2.10.1
    # via pydantic
pyparsing==3.1.2
    # via httplib2
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via huggingface-hub
requests==2.32.3
    # via
    #   google-api-core
    #   huggingface-hub
rsa==4.9
    # via google-auth
sniffio==1.3.0
    # via
    #   anthropic
    #   anyio
    #   httpx
tokenizers==0.19.1
    # via anthropic
tqdm==4.66.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   openai
typing-extensions==4.8.0
    # via
    #   anthropic
    #   google-generativeai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via requests
